fix: Linux build — enum/non-enum ternary warning and BuildKit stale context
Two issues fixed:

1. GCC -Wextra flags mixed enum/uint8_t in ternary expressions involving
   MLD_IN_SPACE. Added static_cast<uint8_t> to both branches in
   multiplicity_descriptor_engine.h and the canonicalisation test.

2. Documented container CLI v0.9 BuildKit stale context workaround:
   `container builder delete --force` clears the builder's context cache
   when --no-cache alone doesn't pick up source changes. Added to both
   container-build.sh comments and testing-guide.md §11.

Also adds BOOST_BIND_GLOBAL_PLACEHOLDERS to suppress Boost 1.74 pragma
noise on Ubuntu 22.04.

+++ src/main/game/multiplicity_descriptor_engine.h
@@ -399,7 +399,7 @@ public:
         for (pile::ref tab_ref : ctx.original_tableau_piles) {
             const pile& p = ctx.piles[tab_ref];
             uint8_t space_kind = pile_sym
-                ? MLD_IN_SPACE
+                ? static_cast<uint8_t>(MLD_IN_SPACE)
                 : static_cast<uint8_t>(MLD_IN_SPACE + pile_idx);
             for (pile::size_type i = 0; i < p.size(); i++) {
                 card c = p[i];
